add de libft + fillit avec leaks

+++ libft/ft_strsplit.c
@@ -6,13 +6,13 @@
 /*   By: lyhamrou <lyhamrou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 20:14:49 by lyhamrou          #+#    #+#             */
-/*   Updated: 2018/11/17 16:08:35 by lyhamrou         ###   ########.fr       */
+/*   Updated: 2018/12/12 14:59:25 by lyhamrou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include "includes/libft.h"
 
-static int		words(char const *s, char c)
+static unsigned long	words(char const *s, char c)
 {
 	int			i;
 	int			cpt;
@@ -31,7 +31,7 @@ static int		words(char const *s, char c)
 	return (cpt);
 }
 
-static int		letter(char const *s, int start, char c)
+static unsigned long	letter(char const *s, int start, char c)
 {
 	int			i;
 
@@ -44,11 +44,11 @@ static int		letter(char const *s, int start, char c)
 	return (i);
 }
 
-char			**ft_strsplit(char const *s, char c)
+char					**ft_strsplit(char const *s, char c)
 {
-	char		**wds;
-	int			i;
-	int			x;
+	char			**wds;
+	unsigned int	i;
+	unsigned int	x;
 
 	i = 0;
 	x = 0;
